requirements.txt
paho-mqtt>=2.0.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.0.0
schedule>=1.2.0
mysql-connector-python>=8.0.0
